Compliance-OS--main/requirements.txt
# ══════════════════════════════════════════════════════════════
# ComplianceOS — Production Dependencies
# ══════════════════════════════════════════════════════════════

# ── Core Framework ──
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.3.0

# ── Database ──
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0        # SQLite async driver (dev)
asyncpg>=0.29.0          # PostgreSQL async driver (prod)

# ── Authentication ──
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# ── AI / LLM ──
groq>=0.8.0
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0

# ── Document Processing ──
PyMuPDF>=1.24.0          # PDF extraction (primary)
pdfplumber>=0.11.0       # PDF extraction (fallback)
python-docx>=1.1.0       # DOCX extraction
pytesseract>=0.3.10      # OCR
pdf2image>=1.17.0        # PDF to Image for OCR
Pillow>=10.3.0           # Image processing

# ── Vector Database ──
chromadb>=0.5.0

# ── HTTP / Middleware ──
starlette>=0.37.0
httpx>=0.27.0
resend>=2.1.0            # Email service

# ── Observability ──
python-json-logger>=2.0.7
sentry-sdk[fastapi]>=2.5.0

# ── Billing ──
stripe>=9.10.0

# ── Deployment ──
gunicorn>=22.0.0
